+++ Project0/src/main/java/dao/IAccount.java
@@ -8,8 +8,8 @@ public interface IAccount {
 	public Account getAccount(int a_id);
 	public boolean addAccount(Account a);
 	public boolean deleteAccount(int a_id);
-	public boolean withdraw(int a_id, int amount);
-	public boolean deposit(int a_id, int amount);
+	public boolean withdraw(int a_id, double amount);
+	public boolean deposit(int a_id, double amount);
 	public boolean withdrawAll(int a_id);
 	public List<Account> allAccounts(int b_id);
